+++ src/runtime/moduleobject.cs
@@ -20,6 +20,12 @@ namespace Python.Runtime
         internal IntPtr dict;
         internal BorrowedReference DictRef => new BorrowedReference(dict);
         protected string _namespace;
+        private IntPtr __all__ = IntPtr.Zero;
+
+        // Attributes to be set on the module according to PEP302 and 451
+        // by the import machinery.
+        static readonly HashSet<string> settableAttributes = 
+            new HashSet<string> {"__spec__", "__file__", "__name__", "__path__", "__loader__", "__package__"};
 
         public ModuleObject(string name)
         {
@@ -47,7 +53,7 @@ namespace Python.Runtime
             var dictRef = Runtime.PyObject_GenericGetDict(ObjectReference);
             PythonException.ThrowIfIsNull(dictRef);
             dict = dictRef.DangerousMoveToPointer();
-
+            __all__ = Runtime.PyList_New(0);
             using var pyname = NewReference.DangerousFromPointer(Runtime.PyString_FromString(moduleName));
             using var pyfilename = NewReference.DangerousFromPointer(Runtime.PyString_FromString(filename));
             using var pydocstring = NewReference.DangerousFromPointer(Runtime.PyString_FromString(docstring));
@@ -181,7 +187,23 @@ namespace Python.Runtime
                 {
                     continue;
                 }
-                GetAttribute(name, true);
+
+                if(GetAttribute(name, true) != null)
+                {
+                    // if it's a valid attribute, add it to __all__
+                    var pyname = Runtime.PyString_FromString(name);
+                    try
+                    {
+                        if (Runtime.PyList_Append(new BorrowedReference(__all__), new BorrowedReference(pyname)) != 0)
+                        {
+                            throw PythonException.ThrowLastAsClrException();
+                        }
+                    }
+                    finally
+                    {
+                        Runtime.XDecref(pyname);
+                    }
+                }
             }
         }
 
@@ -263,6 +285,13 @@ namespace Python.Runtime
                 return self.dict;
             }
 
+            if (name == "__all__")
+            {
+                self.LoadNames();
+                Runtime.XIncref(self.__all__);
+                return self.__all__;
+            }
+
             ManagedType attr = null;
 
             try
@@ -318,6 +347,25 @@ namespace Python.Runtime
             }
             this.cache.Clear();
             base.Clear();
+        }
+
+        /// <summary>
+        /// Override the setattr implementation.
+        /// This is needed because the import mechanics need
+        /// to set a few attributes
+        /// </summary>
+        [ForbidPythonThreads]
+        public new static int tp_setattro(IntPtr ob, IntPtr key, IntPtr val)
+        {
+            var managedKey = Runtime.GetManagedString(key);
+            if ((settableAttributes.Contains(managedKey)) || 
+                (ManagedType.GetManagedObject(val)?.GetType() == typeof(ModuleObject)) )
+            {
+                var self = (ModuleObject)ManagedType.GetManagedObject(ob);
+                return Runtime.PyDict_SetItem(self.dict, key, val);
+            }
+
+            return ExtensionType.tp_setattro(ob, key, val);
         }
 
         protected override void OnSave(InterDomainContext context)
@@ -461,6 +509,7 @@ namespace Python.Runtime
         public static Assembly AddReference(string name)
         {
             AssemblyManager.UpdatePath();
+            var origNs = AssemblyManager.GetNamespaces();
             Assembly assembly = null;
             assembly = AssemblyManager.FindLoadedAssembly(name);
             if (assembly == null)
@@ -479,7 +528,16 @@ namespace Python.Runtime
             {
                 throw new FileNotFoundException($"Unable to find assembly '{name}'.");
             }
+            // Classes that are not in a namespace needs an extra nudge to be found.
+            ImportHook.UpdateCLRModuleDict();
 
+            // A bit heavyhanded, but we can't use the AssemblyManager's AssemblyLoadHandler
+            // method because it may be called from other threads, leading to deadlocks
+            // if it is called while Python code is executing.
+            var currNs = AssemblyManager.GetNamespaces().Except(origNs);
+            foreach(var ns in currNs){
+                ImportHook.AddNamespace(ns);
+            }
             return assembly;
         }
 
@@ -525,6 +583,24 @@ namespace Python.Runtime
                 }
             }
             return names;
+        }
+
+        /// <summary>
+        /// Note: This should *not* be called directly.
+        /// The function that get/import a CLR assembly as a python module.
+        /// This function should only be called by the import machinery as seen
+        /// in importhook.cs
+        /// </summary>
+        /// <param name="spec">A ModuleSpec Python object</param>
+        /// <returns>A new reference to the imported module, as a PyObject.</returns>
+        [ModuleFunction]
+        [ForbidPythonThreads]
+        public static ModuleObject _load_clr_module(PyObject spec)
+        {
+            ModuleObject mod = null;
+            using var modname = spec.GetAttr("name");
+            mod = ImportHook.Import(modname.ToString());
+            return mod;
         }
     }
 }